Fixed the installer to work better, install bin/ and give it a real backup option.

+++ install.sh
@@ -1,16 +1,40 @@
 #!/usr/bin/env bash
 # Usage: ./install.sh [--no-bak]
-# --no-bak erases the .vim/ directory backup, if it exists as a directory.
-
-if [ -d ${HOME}/.vim/ ]; then
-    mv ${HOME}/.vim $HOME/.vim.bak;
+# --no-bak erases the ${PWD} directory backup, if it exists as a directory.
+CURRTS=`date +%s`;
+if [ ! -d ${PWD}/backups/ ]; then
+    mkdir ${PWD}/backups/;
 fi;
-for i in _*
+BACKDIR=${PWD}/backups/$CURRTS;
+mkdir ${BACKDIR};
+for d in _*;
+do
+    orig="${HOME}/${d/_/.}"
+    if [ -f ${orig} ]; then
+        cp -v ${orig} ${BACKDIR};
+    fi;
+    if [ -d ${orig} ]; then
+        mv -v ${orig} ${BACKDIR};
+    fi;
+done;
+for i in _*;
 do 
     source="${PWD}/$i"
     target="${HOME}/${i/_/.}"
-    ln -fivs ${source} ${target}
-done
-if [ "$1" = "--no-bak" ]; then
-    rm -ir ${HOME}/.vim.bak
+    if [ ! -L $target ]; then
+        ln -fivs ${source} ${target};
+    fi;
+done;
+echo "bin/ directory's contents: ";
+ls ${PWD}/bin
+echo "Install the bin/ directory contents to $HOME/bin (y/n)?: "
+read -n 1 doinstbin
+if [ "$doinstbin" == "y" ]; then
+    for b in `ls {$PWD}/bin/`;
+    do
+        cp -v ${PWD}/bin/$b $HOME/bin/;
+    done;
+fi;
+if [ "$1" == "--no-bak" ]; then
+    rm -ir ${BACKDIR}
 fi;
